+++ src/main/java/org/vicky/starterkits/logic/GiveSelectorMode.java
@@ -3,5 +3,7 @@ package org.vicky.starterkits.logic;
 public enum GiveSelectorMode {
     ONCE,
     ALWAYS,
+    ON_DEATH,
+    ON_DEATH_NOT_CLAIMED,
     NONE
 }